// app/lib/theme/useTheme.ts
import { Theme as ReactNavigationTheme } from "@react-navigation/native";
import { create } from "zustand";

import { COLORS } from "./colors";
import { FONT_FAMILIES } from "./families";
import {
  BORDER_RADIUS,
  FONT_SIZES,
  ICON_SIZES,
  LINE_HEIGHTS,
  SPACING,
} from "./sizes";
import { ThemeColors, ThemeState } from "./types";
import { useColorScheme } from "react-native";

const useThemeStore = create<ThemeState>((set) => ({
  theme: "system",
  setTheme: (theme) => set({ theme }),
}));

const getReactNavigationTheme = (
  themeColors: ThemeColors,
  isDark: boolean,
): ReactNavigationTheme => {
  return {
    dark: isDark,
    colors: {
      primary: themeColors.primary as string,
      background: themeColors.background as string,
      card: themeColors.foreground as string,
      text: themeColors.text as string,
      border: themeColors.mist as string,
      notification: themeColors.primary as string,
    },
    fonts: {
      regular: {
        fontWeight: "500",
        fontFamily: FONT_FAMILIES.accent.regular ?? "",
      },
      medium: {
        fontWeight: "500",
        fontFamily: FONT_FAMILIES.accent.medium ?? "",
      },
      bold: {
        fontWeight: "500",
        fontFamily: FONT_FAMILIES.accent.bold ?? "",
      },
      heavy: {
        fontWeight: "500",
        fontFamily: FONT_FAMILIES.accent.extra_bold ?? "",
      },
    },
  };
};

export const useTheme = () => {
  const theme = useThemeStore((state) => state.theme);
  const setTheme = useThemeStore((state) => state.setTheme);
  const getAnyTheme = (requestedTheme: "light" | "dark") => {
    return {
      colors: COLORS[requestedTheme],
      family: FONT_FAMILIES,
      fontSize: FONT_SIZES,
      lineHeight: LINE_HEIGHTS,
      spacing: SPACING,
      borderRadius: BORDER_RADIUS,
      iconSize: ICON_SIZES,
    };
  };
  // const systemTheme = useColorScheme() ?? "light";
  const systemTheme = "dark";

  const activeTheme = theme === "system" ? systemTheme : theme;
  const themeColors = COLORS[activeTheme];
  const reactNavigationTheme = getReactNavigationTheme(
    themeColors,
    activeTheme === "light",
  );

  return {
    theme: {
      colors: themeColors,
      family: FONT_FAMILIES,
      fontSize: FONT_SIZES,
      lineHeight: LINE_HEIGHTS,
      spacing: SPACING,
      borderRadius: BORDER_RADIUS,
      iconSize: ICON_SIZES,
    },
    setTheme,
    activeThemeName: activeTheme,
    inverseActiveThemeName: activeTheme === "light" ? "dark" : "light",
    reactNavigationTheme,
    getAnyTheme,
  };
};
